requirements.txt
pandas
pyarrow
requests
sqlalchemy
psycopg2-binary
python-dotenv
dbt-core
dbt-postgres
streamlit
plotly
